fraude/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
# Dependencias adicionales requeridas
python-dotenv
joblib